- created folders 'src/js/model', 'src/js/service' - moved model, shelly, ip-v4-address, model-service and shelly-service to the new folders - changed some js filenames into camelCase - corrected some linter issues by: replacing modelService's for loop into a while loop, making shellyService's methods static and not a singleton - limited linter to folder 'src/js'
