+++ package.json
@@ -5,7 +5,7 @@
   "main": "app.js",
   "private": true,
   "scripts": {
-    "lint": "eslint --fix --color --format table src",
+    "lint": "eslint --fix --color --format table src/js",
     "test": "",
     "document": "jsdoc --configure jsdoc.conf.json",
     "start": "node src/app.js"
